reorganise
Split modules into a 1lab folder and a non-1lab folder, because 1lab
conflicts with everything.

I wanted to keep the Everything files hidden in _build, but this is hard
because of the way Agda uses `agda-lib` files:
- first, it looks for an `agda-lib` file in the "project root" (by
  default, the current directory, but this can be changed with
  setCommandLineOptions') and uses it to set the default libraries and
  populate include directories;
- then, for each file to be type-checked, it looks for an `agda-lib`
  file as described in https://agda.readthedocs.io/en/latest/tools/package-system.html#the-agda-lib-files-associated-to-a-given-agda-file
  and uses that one for *flags* only.

Also sort the index alphabetically and check for dead links (or rather
agda libraries without a link target).